requirements.txt
# Backend Dependencies
fastapi==0.109.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.48
pymysql==1.1.1
geoalchemy2==0.13.1
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==24.1.0

# Frontend Dependencies (NiceGUI)
nicegui==1.4.37
requests==2.32.3
httpx==0.27.2

# AI Dependencies
google-generativeai==0.8.6
